proper backspace wrapping and mid-buffer returns
previously when backspacing at the beginning of a line when the buffer was not empty, it was not moving the buffer to the end of the previous line. similarly, pressing return key in the middle of a buffer was not copying (partial) contents and moving it into the next (new) line and shifting all the other contents down.