highlight text, see desc
can now highlight text, but the bottom and top rows are buggy when scrolling. see TODO notes in screen_buffer._highlight_text function. Next we need to add copy/paste which should be easy enough - just need to add a (static) buffer in the cursor and reset + update it each time when highlighting text